+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 3.4.0
+
+### Added
+
+- Confirmation `<Popover />`
+- Submit button has been wrapped with it
+
 ## 3.3.5
 
 ### Fixed